Explicitly load the site collection term group
The previous change exclude site collection term groups from being
loaded by the token parser and the object handler for term groups when
specified. This introduced an issue where "termsetid" token, which
reference a site-specific term set, could no longer be resolved because
the group was excluded.

To remedy this, the site collection term group now gets explicitly
processed in both the token parser and the object handler. The token
parser will add additional "termsetid" token if required and the object
handler includes the term group again in the list groups to process.