Use library approach for document libraries when creating folder
"CreateFolderImplementation" now also checks the base type of the parent
list to determine if the library approach should be used when creating a
folder. Previously a folder was always created under the assumption that
the target list is not a library even if this was not the case.

(cherry picked from commit 2ea745942fb226c25403fb497e8b2dbba76e89bc)